Activity formattable content: Code organization
