+++ WordPress/WordPressTest/ActivityLogFormattableContentTests.swift
@@ -33,7 +33,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
     func testPingbackContentIsParsedCorrectly() {
         let dictionary = testData.getPingbackDictionary()
 
-        let pingbackContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let pingbackContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(pingbackContent.count, 1)
 
@@ -47,7 +47,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
 
     func testPostContentIsParsedCorrectly() {
         let dictionary = testData.getPostContentDictionary()
-        let postContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let postContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(postContent.count, 1)
 
@@ -60,7 +60,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
 
     func testCommentContentIsParsedCorrectly() {
         let dictionary = testData.getCommentContentDictionary()
-        let commentContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let commentContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(commentContent.count, 1)
 
@@ -74,7 +74,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
 
     func testThemeContentIsParsedCorrectly() {
         let dictionary = testData.getThemeContentDictionary()
-        let themeContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let themeContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(themeContent.count, 1)
 
@@ -87,7 +87,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
 
     func testSettingContentIsParsedCorrectly() {
         let dictionary = testData.getSettingsContentDictionary()
-        let settingsContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let settingsContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(settingsContent.count, 1)
 
@@ -101,7 +101,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
 
     func testSiteContentIsParsedCorreclty() {
         let dictionary = testData.getSiteContentDictionary()
-        let siteContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let siteContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(siteContent.count, 1)
 
@@ -114,7 +114,7 @@ final class ActivityLogFormattableContentTests: XCTestCase {
 
     func testPluginContentIsParsedCorreclty() {
         let dictionary = testData.getPluginContentDictionary()
-        let pluginContent = ActivityFormattableContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
+        let pluginContent = ActivityContentFactory.content(from: [dictionary], actionsParser: actionsParser, parent: parent)
 
         XCTAssertEqual(pluginContent.count, 1)
